% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{emdatr}
\alias{emdatr}
\alias{emdatr-package}
\title{Global Disaster Losses from the EMDAT database}
\description{
Global Disaster Losses from the EMDAT database
}
\details{
The EMDAT database provides information on human and financial losses
from more than 21,000 major natural global disasters from 1900 to the
present. This package provides the entire data available from EMDAT
and also additional relevant data. Raw data from EMDAT was downloaded
via the web (\url{http://www.emdat.be/database}) and cleaned and enhanced.
\pkg{emdatr} package provides a sample of this cleaned and enhanced data.
Also, \pkg{emdatr} provides functionality to access the entire cleaned
and enhanced EMDAT data.
}

