\docType{package}
\name{emdatr}
\alias{emdatr}
\alias{emdatr-package}
\title{Global Disaster Losses from the EMDAT database}
\description{
  Global Disaster Losses from the EMDAT database
}
\details{
  The EMDAT database provides valuable information on human
  and financial losses from natural disasters around the
  world. The goal of the package is to promote the use of
  EMDAT data, bring transparency to the data, shed light on
  the limitations of the data, and make the analysis of the
  data easier through the R language. Data from EMDAT was
  downloaded via the web
  (\url{http://www.emdat.be/database}) and cleaned and
  enhanced. \pkg{emdatr} package provides a sample of this
  cleaned and enhanced data. Also, \pkg{emdatr} provides
  functionality to access the entire cleaned and enhanced
  EMDAT data.
}

