\docType{data}
\name{usa_cpi}
\alias{usa_cpi}
\title{USA CPI data}
\format{data frame with 100 rows and 2 variables}
\description{
  Consumer Price Index from the US Bureau of Labor
  Statistics, 1913 - 2012
}
\details{
  Variables:

  \itemize{ \item Year - calendar year \item CPI - consumer
  price index }
}
\references{
  Annual average CPI, all urban consumers, obtained from
  the US Bureau of Labor Statistics. Available from
  http://www.bls.gov/cpi/tables.htm
}
\keyword{datasets}

