% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{usa_cpi}
\alias{usa_cpi}
\title{USA CPI data}
\format{data frame with 115 rows and 2 variables}
\usage{
data(usa_cpi)
}
\description{
Consumer Price Index from the US Bureau of Labor Statistics, 1900 - 2014
}
\details{
Variables:

\itemize{
 \item Year - calendar year
 \item CPI - consumer price index
}
}
\references{
Annual average CPI, all urban consumers, obtained from the
US Bureau of  Labor Statistics. Available from \url{http://www.bls.gov/cpi/tables.htm}
}
\keyword{datasets}

